more refactoring of names to avoid confusing
For example topic->name instead of topic->topic. Because the codebase is
getting larger, it gets a bit harder to guess what type something is.

Is 'topic' a topic name (string) or LiveAPITopic pointer?

The new convention is that if something is called 'topic', it is a
pointer or struct. If it is the name, it should be called 'name' or
'topic_name'