Move fixed_data to subfolder, split into tx and rx files.
The code compiles but is completely untested. Most of the logic in
fixed_data_rx is loosely based on the python dummy client. To test it,
the code should request a self-test from the server.

The best way to test is probably to comment-out some of the code that
handles the advanced egde cases, then re-enabling it slowly when the
tests are succeeding.